[SFL-0002] replace BigInteger by Integer

+++ domain/src/main/java/fr/varex13/inputport/BookService.java
@@ -2,9 +2,7 @@ package fr.varex13.inputport;
 
 import fr.varex13.Course;
 
-import java.math.BigInteger;
-
 public interface BookService {
 
-    void handle(Course course, BigInteger duration);
+    void handle(Course course, Integer duration);
 }
